+++ pyflp/channel.py
@@ -1599,7 +1599,6 @@ class ChannelRack(EventModel, ModelCollection[Channel]):
             typ = et.first(ChannelID.Type).value
             groupnum = et.first(ChannelID.GroupNum).value
 
-            # pylint: disable=redefined-outer-name
             ct = Channel  # prevent type error and logic failure below
             if typ == ChannelType.Automation:
                 ct = Automation