+++ tests/test_events.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from pyflp._events import AsciiEvent, DataEventBase, EventEnum, EventTree, U8Event
+from pyflp._events import AsciiEvent, EventEnum, EventTree, U8Event
 from pyflp.exceptions import EventIDOutOfRange, InvalidEventChunkSize
 
 
@@ -12,9 +12,6 @@ def test_id_out_of_range():
 
     with pytest.raises(ValueError):
         AsciiEvent(EventEnum(0), b"1234-decode-me-baby")
-
-    with pytest.raises(EventIDOutOfRange, match="208-255"):
-        DataEventBase(EventEnum(0), b"")
 
 
 def test_invalid_chunk_size():