+++ CHANGELOG.md
@@ -8,6 +8,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2023-05-28
+
+### Changed
+
+- All event parsing happens during `pyflp.parse` itself.
+- `colour.Color` replaced with `pyflp.types.RGBA`.
+
+### Fixed
+
+- Backtracking issues in nested dictionaries.
+
+### Removed
+
+- Bunch of intermediate `EventBase` subclasses.
+
 ## [2.1.1] - 2023-05-24
 
 ### Changed